# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для сервиса справочники",
        "version": "1.0.0"
    },
    "paths": {
        "/api/cities": {
            "get": {
                "tags": [
                    "cities"
                ],
                "summary": "List of cities",
                "description": "Get list of cities",
                "operationId": "getCities",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            }
        },
        "/api/city": {
            "post": {
                "tags": [
                    "cities"
                ],
                "summary": "Create city",
                "description": "Create city",
                "operationId": "createCity",
                "requestBody": {
                    "description": "Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Almaty"
                                    },
                                    "country_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            }
        },
        "/api/city/{id}": {
            "get": {
                "tags": [
                    "cities"
                ],
                "summary": "Get city",
                "description": "Get city",
                "operationId": "getCity",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "city ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            },
            "put": {
                "tags": [
                    "cities"
                ],
                "summary": "Update city",
                "description": "Update city",
                "operationId": "updateCity",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "city ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "country_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            },
            "delete": {
                "tags": [
                    "cities"
                ],
                "summary": "Delete city",
                "description": "Delete city",
                "operationId": "deleteCity",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "city ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            }
        },
...
    },
...
}